+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+**Fixed**
+- using margin width and height in tick and grid generation
+
 ## v1.3.0 (2020-07-30)
 
 **Changed**